working categ and map component, missing link between the two
