+++ package.json
@@ -16,8 +16,9 @@
         "@googlemaps/js-api-loader": "^1.14.3",
         "@googlemaps/react-wrapper": "^1.1.35",
         "@headlessui/react": "^1.4.2",
+        "@mui/icons-material": "^5.10.6",
         "@mui/lab": "^5.0.0-alpha.99",
-        "@mui/material": "^5.10.5",
+        "@mui/material": "^5.10.6",
         "@mui/styled-engine-sc": "^5.10.3",
         "@react-google-maps/api": "^2.12.2",
         "axios": "^0.21.1",
