+++ src/components/DashboardContent/Map.jsx
@@ -2,7 +2,7 @@ import { useMemo } from 'react'
 import { GoogleMap, useLoadScript, Marker } from '@react-google-maps/api'
 
 const mapContainerStyle = {
-    width: '400px',
+    width: '100%',
     height: '400px',
 }
