--- src/components/DashboardContent/TestDashboardContent.jsx
@@ -1,13 +1,0 @@
-import Category from './Category'
-import CategTest from './CategTest'
-import Map from './Map'
-
-export function DashboardContent() {
-    return (
-        <>
-            {/* <Category /> */}
-            {/* <Map /> */}
-            <CategTest />
-        </>
-    )
-}
