+++ src/pages/dashboard.jsx
@@ -1,5 +1,5 @@
+import { DashboardContent } from '@/components/DashboardContent/DashboardContent'
 import { Layout } from '@/components/Layouts/Layout'
-import { DashboardContent } from '@/components/DashboardContent/TestDashboardContent'
 
 export default function Dashboard() {
     return (
